port-blue-ocean-framework/carg/pyproject.toml
[tool.poetry]
name = "port-ocean-carg"
version = "0.1.0"
description = "Port Ocean integration for CARG (Cloud Architecture Resource Graph) system"
authors = ["Richard Agbaje-Dosekun <richard@example.com>"]
readme = "README.md"
homepage = "https://github.com/your-org/port-ocean-carg"
repository = "https://github.com/your-org/port-ocean-carg"
documentation = "https://github.com/your-org/port-ocean-carg#readme"
keywords = ["port", "ocean", "integration", "carg", "devops", "infrastructure"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: System :: Systems Administration",
]
packages = [{include = "port_ocean_carg"}]
include = [
    "CHANGELOG.md",
    "CLIENT_ARCHITECTURE_SUMMARY.md",
    "SETUP_SUMMARY.md",
    ".port/**/*",
]

[tool.poetry.dependencies]
python = "^3.12"
port_ocean = { version = "^0.28.12", extras = ["cli"] }
aiohttp = "^3.9.0"
loguru = "^0.7.0"
pyyaml = "^6.0"
httpx = "^0.28.1"

[tool.poetry.scripts]
carg-extract-json = "port_ocean_carg.extract_port_json:cli_main"
carg-validate = "port_ocean_carg.validate_port_objects:cli_main"
carg-test = "port_ocean_carg.test_integration:cli_main"

[tool.poetry.plugins."port_ocean.integrations"]
carg = "port_ocean_carg.main"

[tool.poetry.group.dev.dependencies]
# Uncomment this if you want to debug the ocean core together with your integration
# port_ocean = { path = '../../', develop = true, extras = ['all'] }
black = "^24.4.2"
mypy = "^1.3.0"
pylint = ">=2.17.4,<4.0.0"
pytest = ">=8.2,<9.0"
pytest-asyncio = ">=0.24.0"
pytest-httpx = ">=0.30.0"
pytest-xdist = "^3.6.1"
ruff = "^0.6.3"
towncrier = "^23.6.0"

[tool.towncrier]
directory = "changelog"
filename = "CHANGELOG.md"
title_format = "## {version} ({project_date})"
underlines = [""]

  [[tool.towncrier.type]]
  directory = "breaking"
  name = "Breaking Changes"
  showcontent = true

  [[tool.towncrier.type]]
  directory = "deprecation"
  name = "Deprecations"
  showcontent = true

  [[tool.towncrier.type]]
  directory = "feature"
  name = "Features"
  showcontent = true

  [[tool.towncrier.type]]
  directory = "improvement"
  name = "Improvements"
  showcontent = true

  [[tool.towncrier.type]]
  directory = "bugfix"
  name = "Bug Fixes"
  showcontent = true

  [[tool.towncrier.type]]
  directory = "doc"
  name = "Improved Documentation"
  showcontent = true

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.mypy]
exclude = [
    'venv',
    '.venv',
]
plugins = [
    "pydantic.mypy"
]

follow_imports = "silent"
warn_redundant_casts = true
warn_unused_ignores = true
disallow_any_generics = true
check_untyped_defs = true
no_implicit_reexport = true

# for strict mypy: (this is the tricky one :-))
disallow_untyped_defs = true


[tool.ruff]
# Never enforce `E501` (line length violations).
ignore = ["E501"]

[tool.pydantic-mypy]
init_forbid_extra = true
init_typed = true
warn_required_dynamic_aliases = true
warn_untyped_fields = true

[tool.black]
line-length = 88
target-version = ['py311']
include = '\.pyi?$'
exclude = '''
/(
  \scripts
  \.toml
  |\.sh
  |\.git
  |\.ini
  |Dockerfile
  |\.venv
)/
'''

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
addopts = "-vv -n auto ./tests"
